requirements.txt
boto3
awslabs-mcp-lambda-handler
